First usable version of the application

+++ app/code/community/Joc/Latch/Helper/Data.php
@@ -29,20 +29,18 @@ class Joc_Latch_Helper_Data extends Mage_Core_Helper_Abstract {
     }
     
     /**
-     * Invoke Latch library for pair account with Latch app
+     * Invoke Latch library for pair a customer account with Latch app
      * 
-     * @param string $token
-     * @param int $user Administrator user object
      * @return array with status and message of the api response
      */
-    public function pair($token, $user = null) {
+    public function pairCustomer($token) {
         $appId = $this->getApplicationId();
         $appSecret = $this->getSecretKey();
         $apiUrl = $this->getApiUrl();
-        
+
         if(!empty($appId) && !empty($appSecret) && !empty($token)){
             require_once(Mage::getBaseDir('lib') . '/Latch/latch.php');
-            
+
             if($apiUrl) {
                 $api = new Latch($appId, $appSecret, $apiUrl);
             } else {
@@ -57,17 +55,13 @@ class Joc_Latch_Helper_Data extends Mage_Core_Helper_Abstract {
             }
 
             if(!empty($accountId)) {
-                if($user) {
-                    /* @var $object Mage_Admin_Model_User */
-                    $object = $user;
-                } else {
-                    /* @var $object Mage_Customer_Model_Customer */
-                    $object = Mage::getSingleton('customer/session')->getCustomer();
-                }
-                
+                /* @var $customer Mage_Customer_Model_Customer */
+                $customer = Mage::getSingleton('customer/session')->getCustomer();
+
                 try {
-                    $object->setData('latch_id', $accountId);
-                    $object->save();
+                    $customer->setData('latch_id', $accountId);
+                    $customer->save();
+
                     return array("status" => 1, "message" => $this->__("Your account was linked with Latch successfully."));
                 } catch (Exception $ex) {
                     return array("status" => 0, "message" => $this->__("Couldn't link the given token with Latch: ") . $this->__($ex->getMessage()));
@@ -83,12 +77,102 @@ class Joc_Latch_Helper_Data extends Mage_Core_Helper_Abstract {
     }
     
     /**
-     * Unpair customer account with Latch
+     * Invoke Latch library for pair an admin account with Latch app
+     * 
+     * @param string $token
+     * @param Mage_Admin_Model_User $user Administrator user object
+     * @return array with status and message of the api response
+     */
+    public function pairAdmin($token, $user = null) {
+        $appId = $this->getApplicationId();
+        $appSecret = $this->getSecretKey();
+        $apiUrl = $this->getApiUrl();
+
+        if(!empty($appId) && !empty($appSecret) && !empty($token)){
+            require_once(Mage::getBaseDir('lib') . '/Latch/latch.php');
+
+            if($apiUrl) {
+                $api = new Latch($appId, $appSecret, $apiUrl);
+            } else {
+                $api = new Latch($appId, $appSecret);
+            }
+
+            $apiResponse = $api->pair($token);
+            $responseData = $apiResponse->getData();
+
+            if(!empty($responseData)) {
+                $accountId = $responseData->{"accountId"};
+            }
+
+            if(!empty($accountId)) {
+                if($user) {
+                    /* @var $user Mage_Admin_Model_User */
+                    $user->setData('latch_id', $accountId);
+                    return array("status" => 1, "message" => $this->__("The account was linked with Latch successfully."));
+                } else {
+                    return array("status" => 0, "message" => $this->__("Can't link non admin user with Latch."));
+                }
+            } elseif($apiResponse->getError() == NULL) {
+                return array("status" => 0, "message" => $this->__("Latch pairing error: Cannot connect to the server. Please, try again later."));
+            } else {
+                return array("status" => 0, "message" => $this->__("Couldn't link the given token with Latch: ") . $this->__($apiResponse->getError()->getMessage()));
+            }
+        } else {
+            return array("status" => 0, "message" => $this->__("Latch pairing error: Invalid parameters."));
+        }
+    }
+    
+    /**
+     * Invoke Latch lib for unpair customer account with Latch
+     * 
+     * @return array with the status of the api response
+     */
+    public function unpairCustomer() {
+        $appId = $this->getApplicationId();
+        $appSecret = $this->getSecretKey();
+        $apiUrl = $this->getApiUrl();
+        
+        if(!empty($appId) && !empty($appSecret)) {
+            require_once(Mage::getBaseDir('lib') . '/Latch/latch.php');
+            
+            if($apiUrl) {
+                $api = new Latch($appId, $appSecret, $apiUrl);
+            } else {
+                $api = new Latch($appId, $appSecret);
+            }
+            
+            $customer = Mage::getSingleton('customer/session')->getCustomer();
+            $latchId = $customer->getData('latch_id');
+            
+            if($latchId) {
+                $apiResponse = $api->unpair($latchId);
+
+                if($apiResponse->getError() == NULL) { 
+                    try {
+                        $customer->setData('latch_id', '');
+                        $customer->save();
+                    } catch (Exception $ex) {
+                        return array("status" => 0, "message" => $this->__("Couldn't unlink with Latch: ") . $this->__($ex->getMessage()));
+                    }
+                    return array("status" => 1, "message" => $this->__("Your account was unlinked with Latch successfully."));
+                } else {
+                    return array("status" => 0, "message" => $this->__("Couldn't unlink your account with Latch: ") . $this->__($apiResponse->getError()->getMessage()));
+                }
+            } else {
+                return array("status" => 0, "message" => $this->__("There is no Latch Id to unlink."));
+            }
+        } else {
+            return array("status" => 0, "message" => $this->__("Your account wasn't unlinked with Latch. Please try again later."));
+        }
+    }
+    
+    /**
+     * Invoke Latch lib for unpair admin account with Latch
      * 
      * @param Mage_Admin_Model_User $user
      * @return array
      */
-    public function unpair($user = null) {
+    public function unpairAdmin($user = null) {
         $appId = $this->getApplicationId();
         $appSecret = $this->getSecretKey();
         $apiUrl = $this->getApiUrl();
@@ -102,28 +186,19 @@ class Joc_Latch_Helper_Data extends Mage_Core_Helper_Abstract {
                 $api = new Latch($appId, $appSecret);
             }
 
-            if($user) {
-                /* @var $object Mage_Admin_Model_User */
-                $object = Mage::getModel('admin/user')->load($user->getId());
-            } else {
-                /* @var $object Mage_Customer_Model_Customer */
-                $object = Mage::getSingleton('customer/session')->getCustomer();
-            }
-
-            $latchId = $object->getData('latch_id');
+            $latchId = Mage::getModel('admin/user')->load($user->getId())->getData('latch_id');
             $apiResponse = $api->unpair($latchId);
             
-            if($apiResponse->getError() == NULL) {                
-                try {
-                    $object->setData('latch_id', '');
-                    $object->save();
-                } catch (Exception $ex) {
-                    return array("status" => 0, "message" => $this->__("Couldn't unlink with Latch: ") . $this->__($ex->getMessage()));
+            if($latchId) {
+                if($apiResponse->getError() == NULL) { 
+                    $user->setData('latch_id', '');
+                    
+                    return array("status" => 1, "message" => $this->__("Your account was unlinked with Latch successfully."));
+                } else {
+                    return array("status" => 0, "message" => $this->__("Couldn't unlink your account with Latch: ") . $this->__($apiResponse->getError()->getMessage()));
                 }
-                
-                return array("status" => 1, "message" => $this->__("Your account was unlinked with Latch successfully."));
             } else {
-                return array("status" => 0, "message" => $this->__("Couldn't unlink your account with Latch: ") . $this->__($apiResponse->getError()->getMessage()));
+                return array("status" => 0, "message" => $this->__("There is no Latch Id to unlink."));
             }
         } else {
             return array("status" => 0, "message" => $this->__("Your account wasn't unlinked with Latch. Please try again later."));
@@ -131,7 +206,7 @@ class Joc_Latch_Helper_Data extends Mage_Core_Helper_Abstract {
     }
     
     /**
-     * Check if user has Latch enabled
+     * Check if customer has Latch enabled
      * 
      * @param string $latchId
      * @param int $userId
@@ -165,6 +240,57 @@ class Joc_Latch_Helper_Data extends Mage_Core_Helper_Abstract {
                         
                         try{
                             $customer->save();
+                        } catch (Exception $ex) {
+                            return array("status" => 0, "message" => $this->__("Something was wrong, please try to log in again later: ") . $this->__($ex->getMessage()));
+                        }
+                    } else {
+                        return array("status" => 0, "message" => $this->__("Something was wrong, please try to log in again later."));
+                    }
+                }
+            }
+            
+            if (!empty($responseData) && $responseData->{"operations"}->{$appId}->{"status"} === "on") {
+                return array("status" => 0, "message" => "");
+            } else {
+                return array("status" => 1, "message" => "Invalid email or password");
+            }
+        }
+    }
+    
+    /**
+     * Check if admin has Latch enabled
+     * 
+     * @param string $latchId
+     * @param Mage_Admin_Model_User $user
+     * @return array
+     */
+    public function getIfAdminLatchEnabled($latchId, $user) {
+        $appId = $this->getApplicationId();
+        $appSecret = $this->getSecretKey();
+        $apiUrl = $this->getApiUrl();
+        
+        if(!empty($latchId) && !empty($appId) && !empty($appSecret)) {
+            require_once(Mage::getBaseDir('lib') . '/Latch/latch.php');
+            
+            if($apiUrl) {
+                $api = new Latch($appId, $appSecret, $apiUrl);
+            } else {
+                $api = new Latch($appId, $appSecret);
+            }
+
+            $apiResponse = $api->status($latchId);
+            $responseData = $apiResponse->getData();
+            $responseError = $apiResponse->getError();
+            
+            if (empty($apiResponse) || (empty($responseData) && empty($responseError))) {
+                return array("status" => 0, "message" => $this->__("Latch is not ready. Please try to log out and log in again."));
+            } else {
+                if (!empty($responseError)) {
+                    if ($responseError->getCode() == 201) {
+                        $user->setData('latch_id', $latchId);
+                        
+                        try{
+                            $user->save();
                         } catch (Exception $ex) {
                             return array("status" => 0, "message" => $this->__("Something was wrong, please try to log in again later: ") . $this->__($ex->getMessage()));
                         }
